travel/requirements.txt
Django
django-extensions
djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-dotenv